requirements.txt
psycopg2-binary==2.9.7
gunicorn==21.2.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
